requirements.txt
absl-py==2.1.0
argcomplete==3.2.2
avalanche-lib==0.5.0
beautifulsoup4==4.12.3
certifi==2024.6.2
charset-normalizer==3.3.2
clarabel==0.9.0
click==8.1.7
cmeel==0.53.3
contourpy==1.2.1
cvxopt==1.3.2
cycler==0.12.1
daqp==0.5.1.post1
dill==0.3.8
docker-pycreds==0.4.0
ecos==2.0.14
filelock==3.15.4
fonttools==4.53.0
fsspec==2024.6.0
gdown==5.2.0
gitdb==4.0.11
GitPython==3.1.43
GPUtil==1.4.0
grpcio==1.64.1
highspy==1.7.1
idna==3.7
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.5
lightning-utilities==0.11.3.post0
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.9.0
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.40
nvidia-nvtx-cu12==12.1.105
osqp==0.6.7.post0
packaging==24.1
pandas==2.2.2
pillow==10.3.0
pipx==1.4.3
piqp==0.4.1
platformdirs==4.2.2
protobuf==4.25.3
proxsuite==0.6.6
psutil==6.0.0
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytorchcv==0.0.67
pytz==2024.1
PyYAML==6.0.1
qdldl==0.1.7.post4
qpalm==1.2.3
qpsolvers==4.3.2
quadprog==0.1.12
requests==2.32.3
scikit-learn==1.5.0
scipy==1.14.0
scs==3.2.4.post3
seaborn==0.13.2
sentry-sdk==2.7.0
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
soupsieve==2.5
sympy==1.12.1
tensorboard==2.17.0
tensorboard-data-server==0.7.2
threadpoolctl==3.5.0
tomli==2.0.1
torch==2.3.1
torchmetrics==1.4.0.post0
torchvision==0.18.1
tqdm==4.66.4
triton==2.3.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
userpath==1.9.1
wandb==0.17.3
Werkzeug==3.0.3
